refactor: add reexports
