+++ src/features/account-details-form/index.ts
@@ -1,0 +1,2 @@
+export * from './account-details.schema';
+export * from './account-details-form';
