+++ src/features/address-display/address-display.tsx
@@ -5,7 +5,7 @@ import { Box } from '@mui/material';
 
 import { Customer } from '@/lib/axios/requests/schemas/customer.schema';
 
-import { AccountDetails } from '../account-details-form/account-details.schema';
+import { AccountDetails } from '../account-details-form';
 import { AddressesContainer } from './addresses-container';
 import { getDefaultAddresses } from './get-default-addresses';
