+++ src/features/password-form/index.ts
@@ -1,0 +1,2 @@
+export * from './password-form';
+export * from './passwords.schema';
